Checkpoint: compile

+++ tempodb/backend/backend.go
@@ -27,7 +27,7 @@ type Reader interface {
 	Tenants() ([]string, error)
 	Blocks(tenantID string) ([]uuid.UUID, error)
 	BlockMeta(blockID uuid.UUID, tenantID string) (*encoding.BlockMeta, error)
-	Bloom(blockID uuid.UUID, tenantID string, bloomShard uint64) ([]byte, error)
+	Bloom(blockID uuid.UUID, tenantID string, bloomShard int) ([]byte, error)
 	Index(blockID uuid.UUID, tenantID string) ([]byte, error)
 	Object(blockID uuid.UUID, tenantID string, offset uint64, buffer []byte) error
